Move driver components into driver directory
